+++ src/Components/Contact/ContactForm.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useFormik } from 'formik';
 import * as Yup from "yup";
-import { postContactValues } from '../../Services/publicApiService';
+import { onSubmitServicePOST } from '../../Services/contactService';
 
 export const ContactForm = () => {
     const [contactValues, setContactValues] = useState({
@@ -37,7 +37,7 @@ export const ContactForm = () => {
         });
     
     const onSubmit = () => {
-        postContactValues(contactValues);
+        onSubmitServicePOST(contactValues);
         resetForm();
     }
