+++ src/Components/Projects/ProjectsForm.scss
@@ -68,9 +68,6 @@ button:disabled {
     }
 }
 @media screen and (min-width: 1024px) {
-    .container {
-        margin-top: 8.89%;
-    }
     #logo {
         display: none;
     }
